test(helpers): share the staged engine home doctor and status each rebuilt
Both files carried the same block — temp $HOME, .cache/kesha/engine/bin, and the
HOME/KESHA_CACHE_DIR/KESHA_ENGINE_BIN triple — as a private helper. They now call one
`stageEngineHome`, which returns the layout and leaves writing the engine to the caller
so a test can still stage a healthy, mute or absent one.
